erss-hwk1-qc58-yz851/docker-deploy/web-app/requirements.txt
Django>3.0
psycopg2
django-multiselectfield
